remove unnecessary casts; from Thomas Pfaff
ok oga@

+++ xutil.c
@@ -161,8 +161,8 @@ xu_setstate(struct client_ctx *cc, int state)
 {
 	long	 dat[2];
 
-	dat[0] = (long)state;
-	dat[1] = (long)None;
+	dat[0] = state;
+	dat[1] = None;
 
 	cc->state = state;
 	XChangeProperty(X_Dpy, cc->win, WM_STATE, WM_STATE, 32,
